replaced render_image with render_image_scale_clip and created macros to reduce verbose without cluttering the call stack

+++ src/interface.h
@@ -26,9 +26,9 @@ void display_fake_loading(unsigned);
 void display_dialog(unsigned);
 void display_quit_dialog(void);
 
-void render_self(ship_t *, short, short);
-void render_foe(ship_t *, short, short);
-void render_bars(ship_t *, SDL_Rect *, short, short, bool);
+void render_self(ship_t *, int, int);
+void render_foe(ship_t *, int, int);
+void render_bars(ship_t *, SDL_Rect *, int, int, bool);
 void render_choices(SDL_Rect *, short, char *[], short, short, short, short);
 void render_belongings(ship_t *);
 void render_combat_box(enum combat_choice, ship_t *);
